Some larger refactoring with macro_rules.
* Renamed `GenericRecord` to `DmapRecord`, and moved into dmap.rs file.
* Moved dmap.rs file containing `Record` trait to src/record.rs
* Finished using macro_rules to simplify lib.rs
* Used macro_rules to simplify io_benchmarking.rs
