+++ benches/io_benchmarking.rs
@@ -1,11 +1,12 @@
 use criterion::{criterion_group, criterion_main, Criterion};
-use dmap::formats::dmap::Record;
+use dmap::record::Record;
 use dmap::formats::fitacf::FitacfRecord;
 use dmap::formats::grid::GridRecord;
 use dmap::formats::iqdat::IqdatRecord;
 use dmap::formats::map::MapRecord;
 use dmap::formats::rawacf::RawacfRecord;
 use dmap::formats::snd::SndRecord;
+use paste::paste;
 use std::fs::File;
 
 fn criterion_benchmark(c: &mut Criterion) {
@@ -15,12 +16,6 @@ fn criterion_benchmark(c: &mut Criterion) {
     c.bench_function("Read GRID", |b| b.iter(|| read_grid()));
     c.bench_function("Read SND", |b| b.iter(|| read_snd()));
     c.bench_function("Read MAP", |b| b.iter(|| read_map()));
-    // c.bench_function("Read Full-size RAWACF", |b| {
-    //     b.iter(|| read_fullsize_rawacf())
-    // });
-    // c.bench_function("Read Full-size FITACF", |b| {
-    //     b.iter(|| read_fullsize_fitacf())
-    // });
 
     // let records = read_iqdat();
     // c.bench_with_input(
@@ -30,53 +25,30 @@ fn criterion_benchmark(c: &mut Criterion) {
     // );
 }
 
-fn read_fitacf() -> Vec<FitacfRecord> {
-    let file = File::open("tests/test_files/test.fitacf").expect("Test file not found");
-    FitacfRecord::read_records(file).unwrap()
+/// Generates benchmark functions for a given DMAP record type.
+macro_rules! read_type {
+    ($type:ident) => {
+        paste! {
+            fn [< read_ $type >]() -> Vec<[< $type:camel Record >]> {
+                let file = File::open(format!("tests/test_files/test.{}", stringify!($type))).expect("Test file not found");
+                [< $type:camel Record >]::read_records(file).unwrap()
+            }
+        }
+    }
 }
 
-fn read_rawacf() -> Vec<RawacfRecord> {
-    let file = File::open("tests/test_files/test.rawacf").expect("Test file not found");
-    RawacfRecord::read_records(file).unwrap()
-}
-
-fn read_fullsize_rawacf() -> Vec<RawacfRecord> {
-    let file = File::open("tests/test_files/20210607.1801.00.cly.a.rawacf.mean")
-        .expect("Test file not found");
-    RawacfRecord::read_records(file).unwrap()
-}
-
-fn read_fullsize_fitacf() -> Vec<FitacfRecord> {
-    let file =
-        File::open("tests/test_files/20210607.1801.00.cly.a.fitacf").expect("Test file not found");
-    FitacfRecord::read_records(file).unwrap()
-}
-
-fn read_iqdat() -> Vec<IqdatRecord> {
-    let file = File::open("tests/test_files/test.iqdat").expect("Test file not found");
-    IqdatRecord::read_records(file).unwrap()
-}
+read_type!(iqdat);
+read_type!(rawacf);
+read_type!(fitacf);
+read_type!(grid);
+read_type!(map);
+read_type!(snd);
 
 // fn write_iqdat(records: &Vec<RawDmapRecord>) {
 //     let file = File::open("tests/test_files/test.iqdat").expect("Test file not found");
 //     dmap::read_records(file).unwrap();
 //     dmap::to_file("tests/test_files/temp.iqdat", records).unwrap();
 // }
-
-fn read_grid() -> Vec<GridRecord> {
-    let file = File::open("tests/test_files/test.grid").expect("Test file not found");
-    GridRecord::read_records(file).unwrap()
-}
-
-fn read_map() -> Vec<MapRecord> {
-    let file = File::open("tests/test_files/test.map").expect("Test file not found");
-    MapRecord::read_records(file).unwrap()
-}
-
-fn read_snd() -> Vec<SndRecord> {
-    let file = File::open("tests/test_files/test.snd").expect("Test file not found");
-    SndRecord::read_records(file).unwrap()
-}
 
 criterion_group! {
     name = benches;
